+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 - [5268](https://github.com/vegaprotocol/vega/issues/5268) - Make validator heartbeat frequency a function of the epoch duration.
 - [5271](https://github.com/vegaprotocol/vega/issues/5271) - Make generated hex IDs lower case
 - [5273](https://github.com/vegaprotocol/vega/issues/5273) - Reward / Transfer to allow payout of reward in an arbitrary asset unrelated to the settlement and by market.
+- [5243](https://github.com/vegaprotocol/vega/issues/5243) - Update equity like share according to spec changes.
 
 ### 🐛 Fixes
 - [4798](https://github.com/vegaprotocol/vega/issues/4978) - Set market pending timestamp to the time at which the market is created.